backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
openai==1.0.0
pdfkit==1.0.0
requests==2.31.0
jinja2==3.1.2
